requirements.txt.txt
Flask
mysql-connector-python
gunicorn  # Necessário para hospedar apps Python